+++ implement/pine/Elm/LanguageServiceInterface/LanguageServiceRequest.cs
@@ -18,13 +18,23 @@ public record FileTreeBlobNode(
 /*
 
 type Request
-    = ProvideHoverRequest ProvideHoverRequestStruct
+    = AddFileRequest (List String) FileTreeBlobNode
+    | DeleteFileRequest (List String)
+    | ProvideHoverRequest ProvideHoverRequestStruct
     | ProvideCompletionItemsRequest ProvideCompletionItemsRequestStruct
     | ProvideDefinitionRequest ProvideHoverRequestStruct
 
  * */
 public abstract record Request
 {
+    public record AddFileRequest(
+        IReadOnlyList<string> FilePath,
+        FileTreeBlobNode Blob)
+        : Request;
+
+    public record DeleteFileRequest(IReadOnlyList<string> FilePath)
+        : Request;
+
     public record ProvideHoverRequest(ProvideHoverRequestStruct Request)
         : Request;
 
@@ -35,15 +45,14 @@ public abstract record Request
 public record ProvideHoverRequestStruct(
     IReadOnlyList<string> FilePathOpenedInEditor,
     int PositionLineNumber,
-    int PositionColumn,
-    string LineText);
+    int PositionColumn);
 
 /*
 
 type alias ProvideCompletionItemsRequestStruct =
     { filePathOpenedInEditor : List String
     , cursorLineNumber : Int
-    , textUntilPosition : String
+    , cursorColumn : Int
     }
 
  * */
@@ -51,7 +60,7 @@ type alias ProvideCompletionItemsRequestStruct =
 public record ProvideCompletionItemsRequestStruct(
     IReadOnlyList<string> FilePathOpenedInEditor,
     int CursorLineNumber,
-    string TextUntilPosition);
+    int CursorColumn);
 
 public static class RequestEncoding
 {
@@ -59,6 +68,40 @@ public static class RequestEncoding
     {
         return request switch
         {
+            Request.AddFileRequest addFileRequest =>
+                ElmValueEncoding.TagAsPineValue(
+                    "AddFileRequest",
+                    [
+                        PineValue.List(
+                            [..addFileRequest.FilePath
+                            .Select(ElmValueEncoding.StringAsPineValue)
+                            ]),
+                        ElmValueEncoding.ElmValueAsPineValue(
+                            new ElmValue.ElmRecord(
+                                [
+                                ("asBase64",
+                                new ElmValue.ElmString(addFileRequest.Blob.AsBase64)),
+
+                                ("asText",
+                                    addFileRequest.Blob.AsText is { } asText
+                                    ?
+                                    ElmValue.TagInstance("Just", [new ElmValue.ElmString(asText)])
+                                    :
+                                    ElmValue.TagInstance("Nothing", [])
+                                    )
+                                ]))
+                    ]),
+
+            Request.DeleteFileRequest deleteFileRequest =>
+                ElmValueEncoding.TagAsPineValue(
+                    "DeleteFileRequest",
+                    [
+                        PineValue.List(
+                            [..deleteFileRequest.FilePath
+                            .Select(ElmValueEncoding.StringAsPineValue)
+                            ])
+                    ]),
+
             Request.ProvideHoverRequest provideHoverRequest =>
                 ElmValueEncoding.TagAsPineValue(
                     "ProvideHoverRequest",
@@ -93,8 +136,6 @@ public static class RequestEncoding
                     PineValueAsInteger.ValueFromSignedInteger(provideHoverRequest.PositionLineNumber)),
                     ("positionColumn",
                     PineValueAsInteger.ValueFromSignedInteger(provideHoverRequest.PositionColumn)),
-                    ("lineText",
-                    ElmValueEncoding.StringAsPineValue(provideHoverRequest.LineText))
                 ]);
     }
 
@@ -110,8 +151,8 @@ public static class RequestEncoding
                         ])),
                     ("cursorLineNumber",
                     PineValueAsInteger.ValueFromSignedInteger(provideCompletionItemsRequest.CursorLineNumber)),
-                    ("textUntilPosition",
-                    ElmValueEncoding.StringAsPineValue(provideCompletionItemsRequest.TextUntilPosition))
+                    ("cursorColumn",
+                    PineValueAsInteger.ValueFromSignedInteger(provideCompletionItemsRequest.CursorColumn))
                 ]);
     }
 }